assignment-7/requirements.txt
pycryptodome>=3.7.1
